Add files via upload
Developed a system to detect fake products using Blockchain Technology using QR code. A registration and login page allows manufacturers and users to sign up and log in, with details stored in an SQL database. All product details are saved in the blockchain with proper timestamp and location details. After logging in, manufacturers can provide product details and generate a QR code, which users can scan to access those details.